refactor: fetcher
